+++ Controls-demo/gridNew/ColumnScroll/DragScrolling/notScrollableCell.wml
@@ -1,4 +1,4 @@
-<ws:partial template="Controls/grid:ColumnTemplate" scope="{{ _options }}" attr:class="js-controls-ColumnScroll__notDraggable">
+<ws:partial template="Controls/grid:ColumnTemplate" scope="{{ _options }}" attr:class="js-controls-DragScroll__notDraggable">
     <ws:contentTemplate>
         <span style="background-color: aliceblue; border-radius: 5px; padding: 2px 4px;">За эту колонку нельзя скроллить</span>
     </ws:contentTemplate>
