--- Controls/_columnScroll/New/ScrollBar/ScrollBar.wml
@@ -1,33 +1,0 @@
-<ws:if data="{{_options.stickyHeader}}">
-    <!-- backgroundStyle transparent служит временным решением ошибки. -->
-    <!-- Если прикладной разработчик не конфигурировал backgroundStyle, он должен быть прозрачным, иначе будет белая подложка у стики шапки. -->
-    <!-- https://online.sbis.ru/opendoc.html?guid=6f6af645-3f36-4081-810c-652fa25eebd5 -->
-    <Controls.scroll:StickyBlock attr:class="controls-ColumnScroll__thumbWrapper {{ _fixedClass }} js-controls-ColumnScroll__thumbWrapper"
-                                  attr:style="z-index: 4;"
-                                  backgroundStyle="{{ (_options.backgroundStyle || _options.style) === 'default' ? 'transparent' : (_options.backgroundStyle || _options.style) }}"
-                                  mode="stackable"
-                                  ntype="scroll"
-                                  fixedZIndex="{{5}}"
-                                  shadowVisibility="hidden">
-        <Controls.scroll:_Scrollbar name="scrollbar"
-                                    attr:class="js-controls-ColumnScroll__thumb controls-ColumnScroll__thumb"
-                                    position="{{_position}}"
-                                    contentSize="{{_contentSize}}"
-                                    direction="horizontal"
-                                    shouldSetMarginTop="{{ true }}"
-                                    trackVisible="{{ true }}"
-                                    on:positionChanged="_onPositionChanged()"
-                                    on:draggingChanged="_onDraggingChanged()"/>
-    </Controls.scroll:StickyBlock>
-</ws:if>
-<ws:else>
-    <Controls.scroll:_Scrollbar name="scrollbar"
-                                attr:class="js-controls-ColumnScroll__thumb controls-ColumnScroll__thumb"
-                                position="{{_position}}"
-                                contentSize="{{_contentSize}}"
-                                direction="horizontal"
-                                shouldSetMarginTop="{{ false }}"
-                                trackVisible="{{ true }}"
-                                on:positionChanged="_onPositionChanged()"
-                                on:draggingChanged="_onDraggingChanged()"/>
-</ws:else>
