+++ Controls/_grid/styles/_Table.less
@@ -10,7 +10,7 @@
     table-layout: auto;
   }
 
-  &.controls-Grid_columnScroll {
+  &.controls-ColumnScroll__content {
     position: static;
   }
 
@@ -187,7 +187,7 @@
     vertical-align: bottom;
   }
 
-  .controls-Grid__cell_fixed, .controls-GridNew__cell_fixed {
+  .controls-GridView__cell_fixed {
     position: relative;
   }
