+++ Controls/_masterDetail/Base/_Base.less
@@ -119,17 +119,17 @@
 }
 
 // Набор классов для установки градиента горизонтального скролла через backgroundStyle (master, detail)
-.controls-horizontal-gradient-master {
+.controls-ColumnScroll__shadow-master {
    background: linear-gradient(to right, var(--columnscroll_shadow-start-color_grid), var(--columnscroll_shadow-end-color_grid));
    border-left: var(--columnscroll_separator_thickness_grid) solid var(--columnscroll_separator_color_grid);
 }
 
-.controls-horizontal-gradient-detail {
+.controls-ColumnScroll__shadow-detail {
    background: linear-gradient(to right, var(--columnscroll_shadow-start-color_grid), var(--columnscroll_shadow-end-color_grid));
    border-left: var(--columnscroll_separator_thickness_grid) solid var(--columnscroll_separator_color_grid);
 }
 
-.controls-horizontal-gradient-detailContrast {
+.controls-ColumnScroll__shadow-detailContrast {
    background: linear-gradient(to right, var(--columnscroll_shadow-start-color_grid), var(--columnscroll_shadow-end-color_grid));
    border-left: var(--columnscroll_separator_thickness_grid) solid var(--columnscroll_separator_color_grid);
 }
