+++ Controls/columnScroll.ts
@@ -3,42 +3,30 @@ import ColumnScrollController, {
     JS_SELECTORS as COLUMN_SCROLL_JS_SELECTORS
 } from 'Controls/_columnScroll/ColumnScrollController';
 
-import NewColumnScrollController, {IControllerOptions as INewColumnScrollControllerOptions} from 'Controls/_columnScroll/New/ColumnScrollController';
-
 import DragScrollController, {
     IDragScrollParams as IDragScrollControllerOptions,
     JS_SELECTORS as DRAG_SCROLL_JS_SELECTORS
 } from 'Controls/_columnScroll/DragScrollController';
 
 import ScrollBar, {IScrollBarOptions} from 'Controls/_columnScroll/ScrollBar/ScrollBar';
-import NewScrollBar, {IScrollBarOptions as INewScrollBarOptions} from 'Controls/_columnScroll/New/ScrollBar/ScrollBar';
 import {isInLeftSwipeRange} from 'Controls/_columnScroll/ColumnScrollUtil';
-import Shadows from 'Controls/_columnScroll/New/Shadows/Shadows';
-import StyleContainers from 'Controls/_columnScroll/New/StyleContainers/StyleContainers';
-import DragScrollOverlay from 'Controls/_columnScroll/New/DragScrollOverlay/DragScrollOverlay';
+import Shadows from 'Controls/_columnScroll/Shadows/Shadows';
+import StyleContainers from 'Controls/_columnScroll/StyleContainers/StyleContainers';
+import DragScrollOverlay from 'Controls/_columnScroll/DragScrollOverlay/DragScrollOverlay';
 
 export {
     ColumnScrollController,
     IColumnScrollControllerOptions,
     COLUMN_SCROLL_JS_SELECTORS,
 
+    DragScrollController,
+    IDragScrollControllerOptions,
     DRAG_SCROLL_JS_SELECTORS,
 
     ScrollBar,
     IScrollBarOptions,
-
-    isInLeftSwipeRange
-};
-
-export {
-    NewColumnScrollController,
-    INewColumnScrollControllerOptions,
-
-    DragScrollController,
-    IDragScrollControllerOptions,
-
-    NewScrollBar, INewScrollBarOptions,
     Shadows,
     StyleContainers,
-    DragScrollOverlay
+    DragScrollOverlay,
+    isInLeftSwipeRange
 };
